Fix censor manager using wrong index
nlu_censor_manager was trying to use a relative sentence index as a
lookup value in DocTree. Since all indices in DocTree are global to the
document, this caused no results to be returned.

Adds a convenience method to SentenceNode called 'find' that returns a
filtered list of word nodes in the sentence matching the given string
argument.

nlu_censor_manager now uses this instead.